move to pykeen.utils to avoid cyclic imports
