+++ src/pykeen/nn/text.py
@@ -14,8 +14,7 @@ from torch import nn
 from torch_max_mem import maximize_memory_utilization
 from tqdm.auto import tqdm
 
-from ..evaluation.evaluator import determine_maximum_batch_size
-from ..utils import get_preferred_device, resolve_device, upgrade_to_sequence
+from ..utils import determine_maximum_batch_size, get_preferred_device, resolve_device, upgrade_to_sequence
 
 if TYPE_CHECKING:
     from .representation import Representation
